get pay modal working (still needs ui work)
